Split sections in `text` segment into 3 separate segments
The linker will align the segments in the image according to their load
alignment, which causes the `.text` section to be at an offset of 32KB
from the start of the image, outside of the range which is scanned for
multiboot headers. Placing the 32KB-aligned data in a different segment
from the `.text` section allows the `.text` section to be placed at a
smaller offset within the image.