Keep Xen from clobbering GS.BASE when loading guest segments
Similar to the code Colin fixed in commit a1a20d0, Xen is clobbering
SVA's GS.BASE value here because it thinks it has full control of both
the "kernel" and "user" GS.BASE slots provided by the hardware.

In stock Xen this is OK because Xen itself doesn't use GS, and this code
ensures that the correct GS.BASE values are put in place for both the
kernel and userspace of the PV guest being context-switched to. However,
in the SVA config, Xen's helper functions wrgsbase(), wrgsshadow(), and
swapgs() actually manipulate fields in Xen's vCPU descriptor instead of
the hardware registers. (The correct user/kernel GS.BASE for the PV
guest is then loaded into the single GS.BASE slot SVA allows Xen to
control when Xen context-switches to the guest.) The result here is that
Xen's inline assembly unknowingly clobbers SVA's GS.BASE and never
actually manages to read the guest's GS.BASE (as it's trying to do here)
to store in the requisite vCPU descriptor field.

We fix this by special-casing the code for the SVA config and having Xen
save and restore SVA's GS.BASE before and after it temporarily loads %gs
to get the hardware to parse the guest's GS descriptor for it. This is a
hack, because we really shouldn't be trusting Xen to use inline assembly
like this (and certainly not to save and restore a pointer that SVA uses
to access critical data), but it's acceptable in our prototype for now.

This will be documented accordingly in our "hacks" file (see
Papers/2020-VEE-XenSVA/hacks.md).