Implement callbacks for providing secure memory frames to SVA
NOTE: this must be built with SVAVMX commit 47cf2be ("Temporarily
disable some checks while porting Xen's VMX support") or newer, or else
SVA will panic when it tries to get secure memory from Xen and finds
that Xen hasn't removed the frames from its direct map.

(That said, such a panic *shouldn't* actually be triggered by the code
in this commit because SVA won't ask Xen for secure memory frames until
it needs them. But it *would* be triggered by the next code I'm *about*
to commit, in which Xen will hand initialization of VMX over to
sva_initvmx(), which allocates a frame of secure memory for the VMXON
region.)