Don't use `l1e_remove_flags` directly on page table entries
Since SVA makes page table entries unwritable to Xen, any update to a
page table entry needs to go through SVA. However, this macro is used in
many places to operate on "intermediate" page table entries that are not
part of an active page table. Therefore, we opt to modify the few places
where this macro is used directly on page table entries rather than the
many where it isn't. Note that this should cause no loss of correctness,
as this macro does not perform an atomic RMW operation.