Use correct `%rsp` value when showing guest stack dumps
`sva_ialloca` will update the guest`s stack pointer, but that update
isn't propagated to Xen until the `copy_regs_from_sva` call at the end
of bounce frame creation. If we crash the domain before that, then the
guest stack dump will use the `%rsp` value from before the `ialloca`,
producing an incomplete stack dump. We correct this by recalculating the
`%rsp` value that Xen set and ensuring that this new value is passed to
the crash dump.