Ensure our stack is aligned on entry from SVA
The SysV ABI requires that the stack be 16-byte aligned immediately
before calling a function (so that the return address is at an offset of
8 from being 16-byte aligned). The compiler will sometimes generate code
that computes pointer values based on that assumption (e.g. replacing `x
+ 8` with `x | 8`). The stack bottom that we were previously giving to
SVA didn't honor this alignment requirement, which resulted in difficult
to diagnose bugs.