Use native target architecture
NB: Note that this may break certain build configurations (probably
cross-compilation and building for any CPU model other than the current
host), and should be behind a guard. However, it is unguarded because it
is required for our modifications to the alternatives mechanism to work,
and those modifications (namely, evaluating alternatives at compile
time) are also unguarded. Ideally, both of these changes would be behing
a guard (such as `CONFIG_COMPILE_TIME_ALTERNATIVES`), but we didn't
create a guard the first time, so this is unguarded for now.

TODO: Create a config option for compile-time alternatives.