Give responsibility for VM entry/exit to SVA/Shade
There's a lot in this commit (unfortunately), but at a high level, it
continues our incremental port of Xen to SVA/Shade by handing over
responsibility for the core of the VM entry/exit process. This includes:

* Replacing Xen's assembly VMX entry/exit loop with C code that uses the
Shade intrinsics to access the VMX hardware features.

* Having Xen create a "dummy" SVA icontext for HVM guests to comport
with the context switching code that Colin already ported to SVA for PV
guests and the (PV) dom0. There's no host ring 3 for HVM guests, so Xen
will never actually do an iret to execute from this icontext, so it's
just a dummy - but SVA needs *something* to be there or it won't let Xen
switch its ring 0 state to an HVM context coming off a PV context.

* Disable Xen's saving/restoring of things like FPU state and SYSCALL
MSRs during VMX entry and exit, since SVA has responsibility for those
and things will break if Xen tries to overwrite them on its own instead
of using SVA intrinsics.

* Making sure that Xen is declaring host-context page table pages to SVA
when creating a host context for an HVM vCPU. (Note that this is for
*host* PTPs, not EPT PTPs. We haven't ported any EPT stuff yet.)

* Using Shade's new sva_getvmfpu() intrinsic to read an HVM vCPU's XSAVE
context instead of trying to read it from Xen's own data structures,
which aren't being maintained any more since SVA is responsible for FPU
state management.