Fix emulation of CPUID XSAVE size reporting since SVA controls XCR0
Xen's CPUID emulation code for reporting the size needed for an XSAVE
operation assumes that an HVM guest's XCR0 remains active when running
in Xen context for that guest - which is true in vanilla Xen, but not in
SVA-Xen. In SVA-Xen, XCR0 is always set to SVA's hardcoded value when
running in VMX root (host) mode, and the guest's XCR0 is loaded
immediately before VM entry (vice versa on exit) by the sva_runvm()
intrinsic.

Thus, if Xen wants to read the current XCR0 for an HVM guest, it needs
to read it from SVA using the sva_getvmreg() intrinsic. Currently, Xen's
entry/exit code (vmx_do_vmentry_sva()) does this immediately after VM
exit and stores the result in current->arch.xcr0, so we can read it from
there in this CPUID emulation code.

Unfortunately, this means Xen can no longer "cheat" here by simply
querying the real hardware's CPUID instruction and passing through the
result to the guest, since a hardware CPUID would incorrectly use
SVA/Xen's XCR0 instead of the guest's to compute the space required
for an XSAVE area, and thus report the wrong value to the guest.

This confusion was causing Linux HVM guests (specifically PVH, but this
should apply to non-PVH HVM guests too) to report an "XSAVE
inconsistency problem" on cross-checking this with the CPUID leaves
reporting the XSAVE space required for individual X-features, and
subsequently crash due to XSAVES writing past the end of a too-small
buffer.

(It just so happened that for our hardware and guest kernel version,
this resulted in a too-small instead of too-big XSAVE buffer.  For
reference, if it's the other way around, Linux will warn about an "XSAVE
consistency problem" but will apparently still work OK - or at least
won't crash under light usage - since allocating *too much* space for an
XSAVE seems to be mostly harmless. I verified this experimentally by way
of testing an earlier, incorrect version of this patch to Xen which
overestimated the size of an XSAVE buffer.)

The fix, as you can see in the patch, is to have Xen emulate these CPUID
leaves' logic in software, determining the size required for each
feature enabled in the guest's XCR0 based on the static CPUID subleaves
which report the size of each feature individually. Despite the code
looking more complex than the original, it probably actually runs faster
since it can do it without issuing any hardware CPUIDs (a slow
serializing instruction) on the critical path.