Port HVM/VMX GS Shadow context switching to use Shade intrinsics
This code has been broken ever since we ported the PV side of Xen to
SVA, since Xen's rd/wrgsshadow() functions no longer touch the actual
hardware but instead access PV-specific fields under struct vcpu which
are responsible for multiplexing the PV guest's user- and kernel-space
GS Shadow values to share a single physical GSS value on the hardware
(as the other one is taken by SVA, unlike in vanilla Xen, which doesn't
use GSS at all and leaves both slots for the PV guest to use).

With this commit, we can now successfully boot, use, and shut down a PVH
Linux domU under SVA/Shade when the system is under load and must
context-switch the PVH guest. (Previously, we could successfully run
such a PVH guest but it would hang as soon as Xen tried to
context-switch it, due to this GS Shadow issue. We missed this earlier
because there's not normally enough load on the dom0 to force Xen to
context switch a domU which only utilizes one CPU on a 16-logical-CPU
system.)