Port XSS context switching to use Shade intrinsics
The XSS MSR is the counterpart of XCR0 which controls extended processor
features that are only accessible in supervisor mode. We disabled Xen's
code to context-switch it when porting XSAVE functionality to SVA, but
did not (until now) replace the XSS part of that code with a suitable
replacement utilizing SVA intrinsics.

See corresponding commit in the SVAVMX repository:
https://github.com/jtcriswell/SVAVMX/commit/956808d775aa2f9aec9e6b17bf1cf53281c89c83