Don't write directly to SYSCALL MSRs in WRMSR exit handlers
SVA handles context switching of these MSRs in sva_runvm(). We already
disabled Xen's code for this in its context-switching code, but we need
to do the same for its WRMSR handling code to prevent it from clobbering
SVA's syscall handler on the host.

We haven't seen this cause any crashes in practice *yet*, but we've
likely just "gotten lucky", since writes to these MSRs are rare
(normally done just once per vCPU during bootup of a guest) and thus
statistically unlikely to coincide with a context switch from an HVM
vCPU onto a PV one. (That's the only situation in which this would cause
a crash in practice, since SVA restores the correct host values of these
MSRs on every HVM VM exit in sva_runvm(), thus "un-clobbering" the bad
value that Xen wrote here.)