Fix incorrectly ported xsave_enabled()
Previously, I had misunderstood this function to be reporting whether
the *host* is utilizing XSAVE, not whether a particular vCPU is. I
therefore incorrectly modified the function to unconditionally return
true, thinking that the most straightforward way to port it to SVA.

(Prior to this "fix", it would crash on the assert that xsave_area is
non-null, since xsave_area is not used under CONFIG_SVA; and I was
additionally concerned that the xcr0_accum field would not behave
correctly under SVA, which made me keen to avoid using it here by
resorting to this "shortcut". I've since gained a better understanding
of how Xen handles XSAVE and XCR0, and realized that xcr0_accum should,
in fact, operate just as well under SVA as under vanilla Xen - allowing
the much less disruptive, and actually correct, fix of simply disabling
the offending assertion.)

The incorrectness of unconditionally returning true from this function
becomes apparent when one examines Xen's code in
hvm_save_cpu_xsave_states(), expand_xsave_states(), and
xstate_ctxt_size(). expand_xsave_states() relies on a "size" parameter
provided by its caller (hvm_save_cpu_xsave_states()). It makes the
implicit assumption that this size will be at least 576 B when it
subtracts XSTATE_AREA_MIN_SIZE (576 B) from it as a parameter passed to
memset(). (The Xen developers really should have made that precondition
explicit in the comment above expand_xsave_states(), as they did for
other related preconditions.) This is not true when xsave_enabled()
incorrectly tells hvm_save_cpu_xsave_states() that the vCPU has used
XSAVE when it actually hasn't. In such a case, xcr0_accum is 0 - a value
which would cause a #GP if actually loaded into XCR0 - and thus
xstate_ctxt_size() returns the meaningless value of 0 when asked to
determine the size required by an XSAVE area for such a setting.
hvm_save_cpu_xsave_states() will correctly ignore this value when
xsave_enabled() returns false, which will always be true when xcr0_accum
is 0 - or at least, it would be if I hadn't incorrectly hacked up
xsave_enabled() to always return true. :-) Thinking the result of
xstate_ctxt_size() is thus meaningful, hvm_save_cpu_xsave_states()
proceeds to call expand_xsave_states() with a size parameter of 0,
leading to unsigned integer underflow when expand_xsave_states()
proceeds to subtract 576 from that 0. A size of (4 GB - 576 B) is thus
passed to memset(), which proceeds to trash gobs of out-of-bounds
memory.

It's actually a wonder that we got away with this for so long without
crashing. This bug only recently became apparent as I was porting
VMPTRLD and VMCLEAR to use Shade intrinsics, which seems like a
relatively minor change from Xen's perspective (merely replacing some
inline assembly with intrinsic calls) - but apparently shifted the
memory layout enough that the ~4 GB of memory trashed by this bug
overlapped with a read-only region of the direct map, causing the
flamboyant crash we deserved all along.