Port VPID usage and VPID/EPT TLB flushes to use Shade intrinsics
Xen no longer needs to mess with the VPID VMCS field directly, as Shade
is now responsible for it.

Note that this entails a substantial simplification of Xen's VPID/ASID
logic. The straightforward way to assign VPIDs (which Shade uses) is to
simply assign one to each vCPU and leave that assignment intact for the
life of the vCPU. Since Shade already assigns a unique numeric
identifier to each vCPU, we simply set the VPID equal to that.

Vanilla Xen does something more complicated where it draws unused VPIDs
from an incrementing pool. This allows it to implement single-context
VPID flushes (i.e., flushing all TLB entries associated with a
particular vCPU) by simply switching that vCPU's VPID for a fresh one,
thereby disassociating it from its previous translations in the TLB.
Since VPIDs are only 16 bits and would run out rather quickly when used
this way, Xen implements a "generation"/epoch scheme whereby it can
safely reset the incrementing counter by performing an all-contexts
flush.

This is a clever scheme if you want to efficiently perform such
single-context flushes on hardware that only supports all-contexts VPID
flushes (as was the case for some older processors). However, modern
Intel machines, such as our development hardware, support single-context
VPID flushes explicitly with the INVVPID instruction. Not only does the
explicit single-context INVVPID make for simpler code, it eliminates the
need to perform an all-contexts flush whenever the VPID counter
overflows. (That might not happen often enough in practice to be a
serious drag on performance, but at any rate it's better to not have to
do it.)

This commit disables Xen's writes to the VPID field (and the whole
generational-increment scheme) in CONFIG_SVA, in favor of calls to Shade
intrinsics implementing the appropriate types of INVVPID flushes. It
likewise ports Xen's use of INVEPT to use Shade intrinsics, because the
instructions are extremely similar and it made sense to do them
together.