Track SVA VMIDs directly instead of VMCS physical addresses
This detail of the port has been in an awkward state of limbo for quite
a while. Originally, as we gradually handed over more aspects of VMCS
responsibility to Shade intrinsics, we needed Xen to retain access to
the raw physical address of the VMCS so it could perform some operations
directly. By now, we have replaced all such operations with calls to
Shade intrinsics.

This allows us to eliminate the use of two ugly "hack" intrinsics in
Shade which previously allowed Xen to convert between SVA VMIDs and VMCS
addresses.