Ensure VPID can actually be enabled in CONFIG_SVA
The control bit to actually enable the VPID feature was erroneously
being left off since we disabled the code in vmx_vmenter_helper() that
determined on each entry and exit whether VPID should be enabled. It
turns out that Xen proactively disables the VPID bit during VMCS
construction with the intent that it will be re-enabled in
vmx_vmenter_helper() on hardware that supports it.

With this change, VPID will correctly be enabled during VMCS
construction (and left enabled thereafter) as long as the hardware
supports it. (If the hardware doesn't support it, we have bigger
problems, as Shade assumes it does.)