Remove parameters from sva_allocvm() call to reflect new interface
As recently committed to SVAVMX, we no longer take explicit initial
values for VMCS controls and guest register state at VM-allocation time.
That model didn't fit well with Xen's code structure and we had been
hacking around it from the beginning of the port (Xen was passing null
pointers and SVA was hacked to ignore them). Now SVA initializes these
values to safe (but meaningless) defaults and expects Xen to overwrite
them via sva_writevmcs(), sva_setvmreg(), etc. before the first VM
entry.

This matches vanilla Xen's behavior, which is to initialize these fields
to safe defaults and wait for the toolstack in the dom0 to set them to
meaningful values via hypercalls before the nascent vCPU is first
unpaused and allowed to run.