Don't use -no-integrated-as when building for SVA
Xen cautiously forces -no-integrated-as when it detects Clang is the
compiler as a workaround for some assembler bugs in older Clang
versions. SVA's compiler is based on LLVM/Clang 10.0.1 and is new enough
to not be affected by these issues.

In general, building for SVA with -no-integrated-as is not recommended,
since the system assembler is not guaranteed to be able to understand
the bundle-alignment directives issued by SVA's CFI implementation.

This commit eliminates the need for the out-of-tree hacks we've been
using to strip -no-integrated-as from Clang command lines when building
Xen for SVA.